Move out get_auto_mounts and related things from function_utils
It is no longer tied to FunctionInfo in any way
