+++ modal/mount.py
@@ -4,6 +4,9 @@ import asyncio
 import concurrent.futures
 import dataclasses
 import os
+import site
+import sys
+import sysconfig
 import time
 import typing
 from pathlib import Path, PurePosixPath
@@ -23,6 +26,7 @@ from ._utils.grpc_utils import retry_transient_errors
 from ._utils.package_utils import get_module_mount_info
 from .client import _Client
 from .config import config, logger
+from .exception import ModuleNotMountable
 from .object import _get_environment_name, _Object
 
 ROOT_DIR: PurePosixPath = PurePosixPath("/root")
@@ -635,3 +639,77 @@ def _get_client_mount():
         return _create_client_mount()
     else:
         return _Mount.from_name(client_mount_name(), namespace=api_pb2.DEPLOYMENT_NAMESPACE_GLOBAL)
+
+
+SYS_PREFIXES = {
+    Path(p)
+    for p in (
+        sys.prefix,
+        sys.base_prefix,
+        sys.exec_prefix,
+        sys.base_exec_prefix,
+        *sysconfig.get_paths().values(),
+        *site.getsitepackages(),
+        site.getusersitepackages(),
+    )
+}
+
+SYS_PREFIXES |= {p.resolve() for p in SYS_PREFIXES}
+
+
+def _is_modal_path(remote_path: PurePosixPath):
+    path_prefix = remote_path.parts[:3]
+    remote_python_paths = [("/", "root"), ("/", "pkg")]
+    for base in remote_python_paths:
+        is_modal_path = path_prefix in [
+            base + ("modal",),
+            base + ("modal_proto",),
+            base + ("modal_version",),
+            base + ("synchronicity",),
+        ]
+        if is_modal_path:
+            return True
+    return False
+
+
+def get_auto_mounts() -> typing.List[_Mount]:
+    """mdmd:hidden
+
+    Auto-mount local modules that have been imported in global scope.
+    This may or may not include the "entrypoint" of the function as well, depending on how modal is invoked
+    Note: sys.modules may change during the iteration
+    """
+    auto_mounts = []
+    top_level_modules = []
+    skip_prefixes = set()
+    for name, module in sorted(sys.modules.items(), key=lambda kv: len(kv[0])):
+        parent = name.rsplit(".")[0]
+        if parent and parent in skip_prefixes:
+            skip_prefixes.add(name)
+            continue
+        skip_prefixes.add(name)
+        top_level_modules.append((name, module))
+
+    for module_name, module in top_level_modules:
+        if module_name.startswith("__"):
+            # skip "built in" modules like __main__ and __mp_main__
+            # the running function's main file should be included anyway
+            continue
+
+        try:
+            # at this point we don't know if the sys.modules module should be mounted or not
+            potential_mount = _Mount.from_local_python_packages(module_name)
+            mount_paths = potential_mount._top_level_paths()
+        except ModuleNotMountable:
+            # this typically happens if the module is a built-in, has binary components or doesn't exist
+            continue
+
+        for local_path, remote_path in mount_paths:
+            # TODO: use is_relative_to once we deprecate Python 3.8
+            if any(str(local_path).startswith(str(p)) for p in SYS_PREFIXES) or _is_modal_path(remote_path):
+                # skip any module that has paths in SYS_PREFIXES, or would overwrite the modal Package in the container
+                break
+        else:
+            auto_mounts.append(potential_mount)
+
+    return auto_mounts
